+++ src/SafeOrbit/Memory/SafeString/IReadOnlySafeString.cs
@@ -1,5 +1,6 @@
 ﻿using System.Threading.Tasks;
 using SafeOrbit.Common;
+using SafeOrbit.Memory.SafeStringServices;
 
 namespace SafeOrbit.Memory
 {
@@ -8,6 +9,10 @@ namespace SafeOrbit.Memory
     /// </summary>
     public interface IReadOnlySafeString: IAsyncEquatable<string>, IAsyncEquatable<IReadOnlySafeString>
     {
+        /// <summary>
+        ///     Gets whether the current <see cref="IReadOnlySafeString"/> is disposed.
+        /// </summary>
+        bool IsDisposed { get; }
         /// <summary>
         ///     Gets whether the current instance holds any characters.
         /// </summary>
@@ -27,6 +32,22 @@ namespace SafeOrbit.Memory
         ///     Reveals the string as decrypted plain bytes.
         /// </summary>
         Task<byte[]> RevealDecryptedBytesAsync();
+
+
+        /// <summary>
+        ///     Reveals the string as a disposable string.
+        /// </summary>
+        /// <example>
+        ///     <code>
+        ///           using(var secret = safeString.RevealDecryptedString())
+        ///           {
+        ///             // Use secret.String here.  While in the 'using' block, the string is accessible
+        ///             // but pinned in memory.  When the 'using' block terminates, the string is zeroed
+        ///             // out for security, and garbage collected as usual.
+        ///           }
+        ///     </code>
+        /// </example>
+        Task<IDisposableString> RevealDecryptedStringAsync();
 
         /// <summary>
         ///     Gets encrypted bytes of a character at given <paramref name="index"/>.