+++ src/SafeOrbit/Memory/SafeString/Marshalling/IDisposableString.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace SafeOrbit.Memory.SafeStringServices
+{
+    public interface IDisposableString : IDisposable
+    {
+        /// <summary>
+        ///     Gets the disposable string.
+        /// </summary>
+        /// <value>The disposable string.</value>
+        string String { get; }
+    }
+}